# packages/contracts/openapi/bifrost.v1.yaml
openapi: 3.1.0
info:
  title: Bifrost API
  version: 0.1.0
paths:
...
  /api/v1/admin/users:
    get:
      tags: [AdminUsers]
      operationId: listAdminUsers
      summary: List users with pagination and filters.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/Status"
        - name: roleId
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "401":
          $ref: "#/components/responses/ApiFailure"
    post:
      tags: [AdminUsers]
      operationId: createAdminUser
      summary: Create a user and assign roles.
      security:
        - BearerAuth: []
      responses:
        "201":
          $ref: "#/components/responses/ApiSuccess"
        "400":
          $ref: "#/components/responses/ApiFailure"
        "409":
          $ref: "#/components/responses/ApiFailure"
  /api/v1/admin/users/{userId}:
    get:
      tags: [AdminUsers]
      operationId: getAdminUser
      summary: Return one user.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "404":
          $ref: "#/components/responses/ApiFailure"
    patch:
      tags: [AdminUsers]
      operationId: updateAdminUser
      summary: Update user profile fields and roles.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "404":
          $ref: "#/components/responses/ApiFailure"
  /api/v1/admin/users/{userId}/service-overrides:
    get:
      tags: [AdminPolicy]
      operationId: listUserServiceOverrides
      summary: List user-level service allow/deny overrides.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
    put:
      tags: [AdminPolicy]
      operationId: replaceUserServiceOverrides
      summary: Replace user-level service allow/deny overrides.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "400":
          $ref: "#/components/responses/ApiFailure"
...
  /api/v1/admin/users/{userId}/status:
    post:
      tags: [AdminUsers]
      operationId: setAdminUserStatus
      summary: Enable or disable a user.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
        "422":
          $ref: "#/components/responses/ApiFailure"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: BifrostAccessToken
    ServiceAccessCookie:
      type: apiKey
      in: cookie
      name: bifrost_access_ticket
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        default: 20
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
    Status:
      name: status
      in: query
      schema:
        type: string
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ApiSuccess:
      description: Unified Bifrost success envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiSuccessEnvelope"
    ApiFailure:
      description: Unified Bifrost failure envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiFailureEnvelope"
  schemas:
    ErrorCode:
      type: string
      enum:
        - COMMON_BAD_REQUEST
        - COMMON_VALIDATION_FAILED
        - COMMON_NOT_FOUND
        - COMMON_CONFLICT
        - COMMON_RATE_LIMITED
        - COMMON_INTERNAL_ERROR
        - AUTH_INVALID_CREDENTIALS
        - AUTH_INVALID_TOKEN
        - AUTH_REFRESH_TOKEN_INVALID
        - AUTH_SESSION_EXPIRED
        - AUTH_SESSION_REVOKED
        - AUTH_PASSWORD_REQUIRED
        - AUTH_PASSWORD_TOO_WEAK
        - USER_DISABLED
        - USER_NOT_FOUND
        - USER_ALREADY_EXISTS
        - USER_CANNOT_DISABLE_SELF
        - USER_CANNOT_DELETE_SELF
        - USER_LAST_ADMIN_REQUIRED
        - ROLE_NOT_FOUND
        - ROLE_ALREADY_EXISTS
        - ROLE_IN_USE
        - ROLE_SYSTEM_ROLE_LOCKED
        - ROLE_POLICY_INVALID
        - DEVICE_NOT_FOUND
        - DEVICE_NOT_TRUSTED
        - DEVICE_DISABLED
        - DEVICE_KEY_INVALID
        - DEVICE_CHALLENGE_EXPIRED
        - DEVICE_ALREADY_BOUND
        - DEVICE_LIMIT_EXCEEDED
        - SERVICE_NOT_FOUND
        - SERVICE_DISABLED
        - SERVICE_ALREADY_EXISTS
        - SERVICE_UPSTREAM_INVALID
        - SERVICE_UPSTREAM_UNREACHABLE
        - SERVICE_ROUTE_INVALID
        - POLICY_ACCESS_DENIED
        - POLICY_USER_DENIED
        - POLICY_ROLE_DENIED
        - POLICY_DEVICE_DENIED
        - POLICY_RULE_INVALID
        - AUDIT_EVENT_NOT_FOUND
        - AUDIT_QUERY_INVALID
        - AUDIT_EXPORT_TOO_LARGE
        - GATEWAY_BAD_UPSTREAM
        - GATEWAY_UPSTREAM_TIMEOUT
        - GATEWAY_ROUTE_NOT_FOUND
        - GATEWAY_REQUEST_TOO_LARGE
...
    Pagination:
      type: object
      additionalProperties: false
      required:
        - page
        - pageSize
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ApiMeta:
      type: object
      additionalProperties: false
      required:
        - requestId
        - timestamp
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        pagination:
          $ref: "#/components/schemas/Pagination"
    ApiError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - userMessage
        - details
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        userMessage:
          type: string
        details:
          type: object
          additionalProperties: true
    ApiSuccessEnvelope:
      type: object
      additionalProperties: false
      required:
        - success
        - data
        - meta
        - error
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          additionalProperties: true
        meta:
          $ref: "#/components/schemas/ApiMeta"
        error:
          type: "null"
    ApiFailureEnvelope:
      type: object
      additionalProperties: false
      required:
        - success
        - data
        - meta
        - error
      properties:
        success:
          type: boolean
          const: false
        data:
          type: "null"
        meta:
          $ref: "#/components/schemas/ApiMeta"
        error:
          $ref: "#/components/schemas/ApiError"